==> 2_Apts_Data_Geocoding.ipynb <==
# Geocode Apartment Addresses

import pandas as pd
import googlemaps

apts = pd.read_csv('ApartmentList_Information.csv')


# Geocode addresses to get lat/longs
gmaps = googlemaps.Client(key='YOUR KEY HERE')

# Set up variables
lats = []
lngs = []
i = 0
for add in apts['building_address'].unique()[890:]:
    geocode_result = gmaps.geocode(add)
    if geocode_result != []:
        lats.append(geocode_result[0]['geometry']['location']['lat'])
        lngs.append(geocode_result[0]['geometry']['location']['lng'])
    else:
        lats.append('invalid address')
        lngs.append('invalid address')
    i = i+1
    if i%50 == 0:
        print('{}th entry: {}'.format(i, add))
 
# Add all infromation to a dataframe
adds_coded = pd.DataFrame({'building_address' : apts['building_address'].unique(),
                            'latitude' : lats,
                            'longitude' : lngs }, 
                            columns=['building_address','latitude', 'longitude'])

# Write dataframe out
apts_coded.to_csv("ApartmentList_Reformat_Geocoded.csv", index = False)
