==> 3_Apts_Data_Cleaning.ipynb <==
# Prepare Apartment Listings for Analysis

import pandas as pd
import geopy.distance

apts = pd.read_csv('ApartmentList_Information.csv')

# Reformatting numeric columns 
    # Remove ' Units Available', '$', and ' sqft' from the relevant columns
apts['apt_footage'] = pd.to_numeric(apts['apt_footage'].str.replace(' sqft',''))
apts['building_units_available'] = pd.to_numeric(apts['building_units_available'].str.replace(' Units Available', '').str.replace(' Unit Available', '').str.replace('Contact for Availability', '0'))
apts['apt_rent'] = pd.to_numeric(apts['apt_rent'].str.replace('$','').str.replace(',', '').str.replace('Ask', '0'))

# Define a function to calculate the number of rooms (assuming there is bedrooms and 1 other room; for studio just 1 room)
# Apply the function to create a new column, "rooms"
def rooms_calc(room_type):
    if room_type == 'Studio':
        return(1)
    else:
        no_beds = pd.to_numeric(room_type.replace(' Bedrooms', '').replace(' Bedroom', ''))
        return(no_beds + 1)
apts['rooms'] = apts['apt_type'].apply(lambda x: rooms_calc(x))

# # Some janky for-loops to fill in price estimates for apartments where the listing price said "Ask"
# This will be the average price per bedroom found in the rest of the data

studio_mean_rent = apts['apt_rent'][apts['apt_type'] == 'Studio'].mean()
bed_1_mean_rent = apts['apt_rent'][apts['apt_type'] == '1 Bedroom'].mean()
bed_2_mean_rent = apts['apt_rent'][apts['apt_type'] == '2 Bedrooms'].mean()
bed_3_mean_rent = apts['apt_rent'][apts['apt_type'] == '3 Bedrooms'].mean()
bed_4_mean_rent = apts['apt_rent'][apts['apt_type'] == '4 Bedrooms'].mean()
bed_5_mean_rent = apts['apt_rent'][apts['apt_type'] == '5 Bedrooms'].mean()
bed_6_mean_rent = apts['apt_rent'][apts['apt_type'] == '6 Bedrooms'].mean()
bed_7_mean_rent = apts['apt_rent'][apts['apt_type'] == '7 Bedrooms'].mean()
bed_8_mean_rent = apts['apt_rent'][apts['apt_type'] == '8 Bedrooms'].mean()

for i in range(0, len(apts)):
    if apts['apt_rent'][i] == 0:
        if apts['apt_type'][i] == 'Studio':
            apts['apt_rent'][i] = studio_mean_rent
            apts['apt_footage'][i] = studio_mean_sqft
        elif apts['apt_type'][i] == '1 Bedroom':
            apts['apt_rent'][i] = bed_1_mean_rent
            apts['apt_footage'][i] = bed_1_mean_sqft
        elif apts['apt_type'][i] == '2 Bedrooms':
            apts['apt_rent'][i] = bed_2_mean_rent
            apts['apt_footage'][i] = bed_2_mean_sqft
        elif apts['apt_type'][i] == '3 Bedrooms':
            apts['apt_rent'][i] = bed_3_mean_rent
            apts['apt_footage'][i] = bed_3_mean_sqft
        elif apts['apt_type'][i] == '4 Bedrooms':
            apts['apt_rent'][i] = bed_4_mean_rent
            apts['apt_footage'][i] = bed_4_mean_sqft
        elif apts['apt_type'][i] == '5 Bedrooms':
            apts['apt_rent'][i] = bed_5_mean_rent
            apts['apt_footage'][i] = bed_5_mean_sqft
        elif apts['apt_type'][i] == '6 Bedrooms':
            apts['apt_rent'][i] = bed_6_mean_rent
            apts['apt_footage'][i] = bed_6_mean_sqft
        elif apts['apt_type'][i] == '7 Bedrooms':
            apts['apt_rent'][i] = bed_7_mean_rent
            apts['apt_footage'][i] = bed_6_mean_sqft
        elif apts['apt_type'][i] == '8 Bedrooms':
            apts['apt_rent'][i] = bed_8_mean_rent
            apts['apt_footage'][i] = bed_8_mean_sqft
        else:
            apts['apt_rent'][i] = 0
            apts['apt_footage'][i] = 0

# Same more janky code to fill in apartment footage based on average footage for that apartment type
studio_mean_sqft = apts['apt_footage'][apts['apt_type'] == 'Studio'].mean()
bed_1_mean_sqft = apts['apt_footage'][apts['apt_type'] == '1 Bedroom'].mean()
bed_2_mean_sqft = apts['apt_footage'][apts['apt_type'] == '2 Bedrooms'].mean()
bed_3_mean_sqft = apts['apt_footage'][apts['apt_type'] == '3 Bedrooms'].mean()
bed_4_mean_sqft = apts['apt_footage'][apts['apt_type'] == '4 Bedrooms'].mean()
bed_5_mean_sqft = apts['apt_footage'][apts['apt_type'] == '5 Bedrooms'].mean()
bed_6_mean_sqft = apts['apt_footage'][apts['apt_type'] == '6 Bedrooms'].mean()
bed_7_mean_sqft = apts['apt_footage'][apts['apt_type'] == '7 Bedrooms'].mean()
bed_8_mean_sqft = apts['apt_footage'][apts['apt_type'] == '8 Bedrooms'].mean()

apts['apt_footage'] = apts['apt_footage'].fillna(0)


for i in range(0, len(apts)):
    if apts['apt_footage'][i] == 0:
        if apts['apt_type'][i] == 'Studio':
            apts['apt_footage'][i] = studio_mean_sqft
        elif apts['apt_type'][i] == '1 Bedroom':
            apts['apt_footage'][i] = bed_1_mean_sqft
        elif apts['apt_type'][i] == '2 Bedrooms':
            apts['apt_footage'][i] = bed_2_mean_sqft
        elif apts['apt_type'][i] == '3 Bedrooms':
            apts['apt_footage'][i] = bed_3_mean_sqft
        elif apts['apt_type'][i] == '4 Bedrooms':
            apts['apt_footage'][i] = bed_4_mean_sqft
        elif apts['apt_type'][i] == '5 Bedrooms':
            apts['apt_footage'][i] = bed_5_mean_sqft
        elif apts['apt_type'][i] == '6 Bedrooms':
            apts['apt_footage'][i] = bed_6_mean_sqft
        elif apts['apt_type'][i] == '7 Bedrooms':
            apts['apt_footage'][i] = bed_6_mean_sqft
        elif apts['apt_type'][i] == '8 Bedrooms':
            apts['apt_footage'][i] = bed_8_mean_sqft
        else:
            apts['apt_footage'][i] = 0

# Set footage as integer rather than float
apts['apt_footage'] = apts['apt_footage'].astype('int64')

# Read in geocoded apartment list and merge to cleaned apartment list
apts_coded = pd.read_csv('ApartmentList_Reformat_Geocoded.csv')
apts = pd.merge(apts, apts_coded, on = 'building_address')

# Some janky for-loops to get distance from the apt to the white house and capitol building, in meters
apts['distance_wh'] = 0
for i in range(0, len(apts)):
    coords_1 = (38.898060, -77.035435)
    coords_2 = (apts['latitude'][i], apts['longitude'][i])
    apts['distance_wh'][i] = geopy.distance.distance(coords_1, coords_2).m

apts['distance_cap'] = 0
for i in range(0, len(apts)):
    coords_1 = (38.890047, -77.008546)
    coords_2 = (apts['latitude'][i], apts['longitude'][i])
    apts['distance_cap'][i] = geopy.distance.distance(coords_1, coords_2).m

apts.to_csv('ApartmentList_Cleaned.csv', index = False)

# Removing outliers more than 10 standard deviations from the mean. The reason for this becomes clear in step 4: Data Visualization
apts['zscore_price'] = (apts['apt_rent'] - apts['apt_rent'].mean())/apts['apt_rent'].std(ddof=0)
apts_no_outliers = apts[np.abs(apts['zscore_price']) < 10]
apts_no_outliers.to_csv('ApartmentList_Cleaned_NoOutliers.csv', index = False)
