==> 7_Apts_Data_Prediction.ipynb <==
# Use Pre-Tuned Model for Prediction

import pandas as pd
import pickle
apts_ml = pd.read_csv('ApartmentList_ML.csv')

rf = pickle.load(open('finalized_rf_model.sav', 'rb'))

# create a new dataframe that is indexed like the trained model
newdata = pd.DataFrame().reindex_like(apts_ml)
newdata.fillna(value=0, inplace=True)

# delete the variable to be predicted
del newdata['apt_rent']
newdata = newdata.iloc[[1]]

# insert information about your apartment 
newdata['rooms'] = 3
newdata['apt_footage'] = 600
newdata['distance_wh'] = 3218
newdata['distance_cap'] = 5632
newdata['building_units_available'] = 2
newdata['adj_open'] = 0
newdata['adj_large'] = 0
newdata['adj_private'] = 0
newdata['adj_great'] = 1
newdata['adj_new'] = 0
newdata['adj_beautiful'] = 0
newdata['adj_high'] = 0
newdata['adj_natural'] = 0
newdata['adj_historic'] = 1
newdata['adj_newly'] = 0
newdata['adj_short'] = 0
newdata['adj_ready'] = 1
newdata['adj_furnished'] = 0
newdata['adj_convenient'] = 0
newdata['adj_windows'] = 0



rf.predict(newdata)
